requirements.txt
# Dexaview Backend – Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
aiomysql==0.2.0          # async MySQL driver
alembic==1.13.3          # database migrations

# Authentication
python-jose[cryptography]==3.3.0   # JWT encoding/decoding
passlib[bcrypt]==1.7.4             # password hashing

# Configuration
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2            # async HTTP client for TestClient
